// src/api/generated/models/updateTaskDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * TodoList API
 * API para gestión de tareas y usuarios
 * OpenAPI spec version: 1.0
 */
import type { UpdateTaskDtoPriority } from "./updateTaskDtoPriority";
import type { UpdateTaskDtoState } from "./updateTaskDtoState";

export interface UpdateTaskDto {
  /**
   * Título de la tarea
   * @minLength 1
   */
  title?: string;
  /** Descripción detallada de la tarea */
  description?: string;
  /** Prioridad de la tarea */
  priority?: UpdateTaskDtoPriority;
  /** ID del usuario asignado a la tarea */
  userId?: string;
  /** Estado inicial de la tarea */
  state?: UpdateTaskDtoState;
}