// src/api/generated/models/updateTaskDtoState.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * TodoList API
 * API para gestión de tareas y usuarios
 * OpenAPI spec version: 1.0
 */

/**
 * Estado inicial de la tarea
 */
export type UpdateTaskDtoState =
  (typeof UpdateTaskDtoState)[keyof typeof UpdateTaskDtoState];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateTaskDtoState = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;